Efficiency improvements
Removed duplicate lines and also only generate points once at the start of the program rather than whenever an update is made. Also fixed a small bug in the 24 cell where I did not get all 24 permutations of the points correctly. Also forgot to rotate the destination point when drawing up the shape.